added tests for user model and helpers

+++ jwt-auth/database/index.ts
@@ -3,50 +3,41 @@ import { MongoMemoryServer } from "mongodb-memory-server";
 
 const { DB_NAME, DB_USER, DB_PASSWORD, DB_HOST, DB_PORT } = process.env;
 
-const mongoMemory = new MongoMemoryServer();
+let mongoMemory: MongoMemoryServer | null = null;
 
-async function connect(useMemoryDb?: "memory") {
-  if (useMemoryDb === "memory") {
-    const uri = mongoMemory.getUri();
-    await mongoose.connect(uri, {});
-  } else {
-    if (
-      DB_HOST === undefined ||
-      DB_PORT === undefined ||
-      DB_NAME === undefined ||
-      DB_USER === undefined ||
-      DB_PASSWORD === undefined
-    ) {
-      throw new Error("Database connection string is not defined");
-    }
-    if (mongoose.connection.readyState >= 1) return;
-    await mongoose
-      .connect(
-        `mongodb://${DB_USER}:${DB_PASSWORD}@${DB_HOST}:${DB_PORT}/${DB_NAME}`
-      )
-      .catch((err) => console.log(err));
+async function connect() {
+  if (
+    DB_HOST === undefined ||
+    DB_PORT === undefined ||
+    DB_NAME === undefined ||
+    DB_USER === undefined ||
+    DB_PASSWORD === undefined
+  ) {
+    throw new Error("Database connection string is not defined");
   }
+  if (mongoose.connection.readyState >= 1) return;
+  await mongoose
+    .connect(
+      `mongodb://${DB_USER}:${DB_PASSWORD}@${DB_HOST}:${DB_PORT}/${DB_NAME}`
+    )
+    .catch((err) => console.log(err));
 }
 
-async function disconnect(useMemoryDb?: "memory") {
-  if (useMemoryDb === "memory") {
-    await mongoMemory.stop();
-  }
+async function disconnect() {
   await mongoose.disconnect().catch((err) => console.log(err));
 }
 
-async function clearMemoryDb() {
-  const collections = mongoose.connection.collections;
-  for (const key in collections) {
-    const collection = collections[key];
-    await collection.deleteMany({});
+async function connectMemoryDb() {
+  mongoMemory = await MongoMemoryServer.create();
+  const uri = mongoMemory.getUri();
+  await mongoose.connect(uri);
+}
+async function closeMemoryDb() {
+  if (mongoMemory) {
+    await mongoose.connection.dropDatabase();
+    await mongoose.connection.close();
+    await mongoMemory.stop();
   }
 }
 
-async function closeMemoryDb() {
-  await mongoose.connection.dropDatabase();
-  await mongoose.connection.close();
-  await mongoMemory.stop();
-}
-
-export default { connect, disconnect, clearMemoryDb, closeMemoryDb };
+export default { connect, disconnect, closeMemoryDb, connectMemoryDb };
